fix: 라이브러리 전역 설정
